Intermediate checkpoint for moving components into routes
