+++ src/routes/[user]/components/AI/GPT.js
