+++ src/routes/[user]/components/AI/text.js
