+++ src/routes/[user]/components/SideBySideView.svelte
