+++ src/routes/[user]/components/Templates/components/IconsDisplay/BasicWhiteboard.svelte
