+++ src/routes/[user]/components/Templates/components/IconsDisplay/ColorPicker.svelte
