+++ src/routes/[user]/components/Templates/store.js
