+++ src/routes/[user]/components/TopNavbar.svelte
@@ -135,4 +135,4 @@
     background-color: rgba(150, 150, 150, 0.1);
     border-bottom: none;
   }
-</style>
+</style>
