+++ src/routes/[user]/components/handleNotifications.js
@@ -1,6 +1,6 @@
 import { getMessaging, getToken, onMessage } from "firebase/messaging";
-import { app } from "../../db/init";
-import User from "../../db/models/User";
+import { app } from "../../../db/init";
+import User from "../../../db/models/User";
 
 export const handleNotificationPermission = (user) => {
   Notification.requestPermission().then((permission) => {
@@ -66,4 +66,4 @@ export const handleSW = async () => {
     }
 
   });
-};
+};
