+++ src/routes/[user]/mobile/FloatingButtonWrapper.svelte
