+++ src/routes/[user]/mobile/Mobile.svelte
@@ -99,9 +99,9 @@
 
 <script>
   import Calendar from '$lib/Calendar/Calendar.svelte'
-  import AI from '$lib/AI/AI.svelte'
-  import ScheduleView from '$lib/Mobile/ScheduleView.svelte'
-  import ListView from '$lib/Mobile/ListView.svelte'
+  import AI from '../components/AI/AI.svelte'
+  import ScheduleView from './ScheduleView.svelte'
+  import ListView from './ListView.svelte'
   import VoiceKeywordDetect from './VoiceKeywordDetect.svelte'
   import TaskPopup from '$lib/TaskPopup/TaskPopup.svelte'
   import FloatingButtonWrapper from './FloatingButtonWrapper.svelte'
