+++ src/routes/[user]/mobile/ScheduleView.svelte
