+++ src/routes/[user]/mobile/VoiceKeywordDetect.svelte
